Fixed the segfaults by not calling copy constructors of nodes when converting from an expression to a function tree.
Changed data for one function from acosh to arccosh. The type names and real
names and SBML names etc. is rather confusing since in some function the
textual representation is different from the type which might be again different
from what it is in SBML.

Improved the wrapper script for the test suite a bit.

+++ semantic-test-suite/copasi_wrapper.sh
@@ -1,22 +1,32 @@
 #!/bin/bash
 
 SBMLFILE=$1
-CONVERTER=/home/sven/copasi/copasi_dev/sbml2copasi
+
+BASEDIR=/home/ralph/workspace/copasi_work/
+
 COPASI_OUTFILE=${SBMLFILE%%.xml}.cps
-ADD_INFO_SCRIPT=/home/sven/copasi/copasi_dev/addTestSuiteInfo.py
+ADD_INFO_SCRIPT=$BASEDIR/semantic-test-suite/addTestSuiteInfo.py
 shift 1
 TASK_ARGS=${*}
-COPASI="/home/sven/copasi/copasi_dev/copasi/CopasiSE/CopasiSE -c ."
+COPASI="$BASEDIR/copasi/CopasiSE/CopasiSE -c ."
 
 #echo $TASK_ARGS
 
 # convert sbml file
-#$CONVERTER $SBMLFILE $COPASI_OUTFILE
-rm $COPASI_OUTFILE
-$COPASI --importSBML $SBMLFILE --save $COPASI_OUTFILE
+if [ -e $COPASI_OUTFILE ];then rm $COPASI_OUTFILE; fi
+echo -e "$COPASI --importSBML $SBMLFILE --save $COPASI_OUTFILE\n"
 
-#add task and report information to copasi file
-$ADD_INFO_SCRIPT $COPASI_OUTFILE $TASK_ARGS
+$COPASI --importSBML $SBMLFILE --save $COPASI_OUTFILE &> /dev/null
 
-# run copasi command line version
-$COPASI $COPASI_OUTFILE
+if [ -e $COPASI_OUTFILE ];then
+  #add task and report information to copasi file
+  echo -e "$ADD_INFO_SCRIPT $COPASI_OUTFILE $TASK_ARGS\n" 
+
+  $ADD_INFO_SCRIPT $COPASI_OUTFILE $TASK_ARGS 
+
+  # run copasi command line version
+  echo -e "$COPASI $COPASI_OUTFILE\n"
+  $COPASI $COPASI_OUTFILE &> /dev/null
+else
+  echo -e "ERROR: Failed to create Copasi file \"$COPASI_OUTFILE\".\n"
+fi
